fix(web): resolve workout notes review follow-ups
